Modify arg order and curry forwardDispatch
This is so that `initializeComponent` can dispatch a component's initial
action. It does this by taking a `(state) => dispatch()` function, which
will most likely be a partially-applied `forwardDispatch`.

Since it's intended that sub-components will be initialised in parent
components' `update`, the root reducer will now pass `store.dispatch`
directly to `update`.